(PA-4331) Fix the logic to only run on older solaris versions

+++ setup/common/003_solaris_cert_fix.rb
@@ -39,16 +39,18 @@ gKDWHrO8Dw9TdSmq6hN35N6MgSGtBxBHEa2HPQfRdbzP82Z+
 EOM
 
 hosts.each do |host|
-  create_remote_file(host, "DigiCertTrustedRootG4.crt.pem", DigiCert)
-  on(host, 'chmod a+r /root/DigiCertTrustedRootG4.crt.pem')
-  on(host, 'cp -p /root/DigiCertTrustedRootG4.crt.pem /etc/certs/CA/')
-  on(host, 'rm /root/DigiCertTrustedRootG4.crt.pem')
-  on(host, '/usr/sbin/svcadm restart /system/ca-certificates')
-  timeout = 60
-  counter = 0
-  while on(host, 'svcs -x ca-certificates').output !~ /State: online/ do
-    raise 'ca-certificates services failed start up' if counter > timeout
-    sleep 5
-    counter = counter + 5
+  if host.platform=~ /solaris-11(\.2)?-(i386|sparc)/
+    create_remote_file(host, "DigiCertTrustedRootG4.crt.pem", DigiCert)
+    on(host, 'chmod a+r /root/DigiCertTrustedRootG4.crt.pem')
+    on(host, 'cp -p /root/DigiCertTrustedRootG4.crt.pem /etc/certs/CA/')
+    on(host, 'rm /root/DigiCertTrustedRootG4.crt.pem')
+    on(host, '/usr/sbin/svcadm restart /system/ca-certificates')
+    timeout = 60
+    counter = 0
+    while on(host, 'svcs -x ca-certificates').output !~ /State: online/ do
+      raise 'ca-certificates services failed start up' if counter > timeout
+      sleep 5
+      counter = counter + 5
+    end
   end
 end
